requirements.txt
rasterio==1.0.21
pylint==2.3.1
keras==2.2.4
tensorflow==1.13.1
matplotlib==3.0.3
scikit-learn==0.19.1
gdal==2.3.3
numpy==1.15.4
pyspark==2.4.1
petastorm==0.7.3
pillow==6.0.0

